chore(selection): cleanup
