--- packages/core/src/components/UserSelection/SelectionBox.tsx
@@ -1,30 +1,0 @@
-import shallow from 'zustand/shallow';
-import { useStore } from '../../hooks/useStore';
-import { ReactFlowState } from '../../types';
-
-const selector = (s: ReactFlowState) => ({
-  userSelectionActive: s.userSelectionActive,
-  userSelectionRect: s.userSelectionRect,
-});
-
-function SelectionBox() {
-  const { userSelectionActive, userSelectionRect } = useStore(selector, shallow);
-  const showSelectionBox = userSelectionActive && userSelectionRect;
-
-  if (!showSelectionBox) {
-    return null;
-  }
-
-  return (
-    <div
-      className="react-flow__selection react-flow__container"
-      style={{
-        width: userSelectionRect.width,
-        height: userSelectionRect.height,
-        transform: `translate(${userSelectionRect.x}px, ${userSelectionRect.y}px)`,
-      }}
-    />
-  );
-}
-
-export default SelectionBox;
